+++ CHANGELOG.md
@@ -1,8 +1,30 @@
-# 6.0.0 (01.09.20)
+# 6.5.0 (03.31.20)
+
 ## Functionality
+
+1. Improved flow for new cards
+2. New dashboard for admins
+3. Can submit for events on the main page
+
+## Behind the hood
+
+1. Packages made current, angular 9, material 9
+2. Removed local storage system
+
+## Future
+
+1. Dashboard view for admins like the current all admin page
+2. Moderate the events via the admin portal (popup)
+3. Export confessions
+
+# 6.0.0 (01.09.20)
+
+## Functionality
+
 Ported to stackblitz
 
 ## Features
+
 1. Event confessions now don't re-require passwords after 2 days if same device.
 2. Changed to a card grid structure
 3. Changed the bottom card actions (likes, comments etc)
@@ -10,18 +32,24 @@ Ported to stackblitz
 5. Reduced the font overall throughout the page
 
 # 3.1.0 (04.08.19)
+
 ## Admin
+
 1. Fixes bug that caused memory leak (excessive function calls)
 
 # 3.0.0 (04.08.19)
+
 ## Admin
+
 1. Fixes no paginator issue
 2. Refactors data password check item
 
 ## Confessions
+
 1. Stores comment counts better
 2. Can sort by basic fields (date/popularity/comments) + ascending/descending
 
 ## MISC
+
 1. New helper functions for data conforming
 2. New array sorter function (modular)